Set SVG background in social section
Used suggestion using css targeting the svg itself vs targeting the div it was in for background color. "Any element that renders in HTML can have a background color" and also got rid of social SVG container div's. This allowed me to change size (prev set to 100% of container div) by setting it based on vh